-Some more refactors
