--- src/main/java/corgitaco/betterweather/helper/IsWeatherForced.java
@@ -1,8 +1,0 @@
-package corgitaco.betterweather.helper;
-
-public interface IsWeatherForced {
-
-    boolean isWeatherForced();
-
-    void setWeatherForced(boolean flag);
-}
