--- src/main/java/corgitaco/betterweather/helper/ViewFrustumGetter.java
@@ -1,8 +1,0 @@
-package corgitaco.betterweather.helper;
-
-import net.minecraft.client.renderer.ViewFrustum;
-
-public interface ViewFrustumGetter {
-
-    ViewFrustum getViewFrustum();
-}
