--- src/main/java/corgitaco/betterweather/helper/WeatherViewFrustum.java
@@ -1,6 +1,0 @@
-package corgitaco.betterweather.helper;
-
-public interface WeatherViewFrustum {
-
-    void forceRenderDistance(int renderDistance, double x, double y, double z);
-}
