+++ src/main/java/corgitaco/betterweather/mixin/entity/MixinBreedGoal.java
@@ -1,6 +1,6 @@
 package corgitaco.betterweather.mixin.entity;
 
-import corgitaco.betterweather.helper.BetterWeatherWorldData;
+import corgitaco.betterweather.helpers.BetterWeatherWorldData;
 import corgitaco.betterweather.season.SeasonContext;
 import net.minecraft.entity.ai.goal.BreedGoal;
 import net.minecraft.entity.passive.AnimalEntity;
