+++ src/main/java/corgitaco/betterweather/mixin/network/MixinPlayerList.java
@@ -1,9 +1,9 @@
 package corgitaco.betterweather.mixin.network;
 
 import com.mojang.authlib.GameProfile;
-import corgitaco.betterweather.helper.BetterWeatherWorldData;
 import corgitaco.betterweather.datastorage.network.NetworkHandler;
 import corgitaco.betterweather.datastorage.network.packet.SeasonPacket;
+import corgitaco.betterweather.helpers.BetterWeatherWorldData;
 import corgitaco.betterweather.season.SeasonContext;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.nbt.CompoundNBT;
