+++ src/main/java/corgitaco/betterweather/server/command/SetSeasonCommand.java
@@ -3,8 +3,8 @@ package corgitaco.betterweather.server.command;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
-import corgitaco.betterweather.helper.BetterWeatherWorldData;
 import corgitaco.betterweather.api.season.Season;
+import corgitaco.betterweather.helpers.BetterWeatherWorldData;
 import corgitaco.betterweather.season.SeasonContext;
 import net.minecraft.command.CommandSource;
 import net.minecraft.command.Commands;
